requirements.txt
pandas==2.0.3
numpy==1.24.3
faiss-cpu==1.7.4
google-generativeai==0.3.1
sentence-transformers==2.2.2
nltk==3.8.1
scikit-learn==1.3.0
sentence-transformers==2.2.2
transformers==4.30.2
huggingface_hub==0.14.1